+++ pkg/testing/templates.go
@@ -20,12 +20,9 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"testing"
 
-	"github.com/google/go-cmp/cmp"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/yaml"
 
@@ -47,216 +44,9 @@ type templateType interface {
 	client.Object
 }
 
-// TemplateTestSuite is a collection of named template tests which may be run together
-type TemplateTestSuite map[string]*TemplateTestCase
-
 type FailedTest struct {
 	name string
 	err  error
-}
-
-// Assert allows testing a TemplateTestSuite when a *testing.T is not available,
-// e.g. when tests are not run from 'go test'
-// It returns a list of the named tests that passed and a list of the named tests that failed with their errors
-func (s *TemplateTestSuite) Assert() ([]string, []*FailedTest) {
-	var (
-		passedTests []string
-		failedTests []*FailedTest
-	)
-
-	testsToRun, _ := s.getTestsToRun()
-
-	for name, testCase := range testsToRun {
-		err := testCase.Run()
-		if err != nil {
-			failedTests = append(failedTests, &FailedTest{name: name, err: err})
-		} else {
-			passedTests = append(passedTests, name)
-		}
-	}
-
-	return passedTests, failedTests
-}
-
-func (s *TemplateTestSuite) HasFocusedTests() bool {
-	_, focused := s.getTestsToRun()
-	return focused
-}
-
-func (s *TemplateTestSuite) Run(t *testing.T) {
-	testsToRun, focused := s.getTestsToRun()
-
-	if focused {
-		defer t.Fatalf("test suite failed due to focused test, check individual test case status")
-	}
-
-	for name, testCase := range testsToRun {
-		tc := testCase
-		t.Run(name, func(t *testing.T) {
-			err := tc.Run()
-			if err != nil {
-				t.Fatal(err)
-			}
-		})
-	}
-}
-
-func (s *TemplateTestSuite) RunConcurrently(t *testing.T) {
-	testsToRun, focused := s.getTestsToRun()
-
-	if focused {
-		defer t.Fatalf("test suite failed due to focused test, check individual test case status")
-	}
-
-	for name, testCase := range testsToRun {
-		tc := testCase
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			err := tc.Run()
-			if err != nil {
-				t.Fatal(err)
-			}
-		})
-	}
-}
-
-func (s *TemplateTestSuite) getTestsToRun() (TemplateTestSuite, bool) {
-	focused := false
-	testsToRun := *s
-	focusedCases := make(map[string]*TemplateTestCase, len(*s))
-
-	for name, testCase := range *s {
-		if testCase.Focus {
-			focusedCases[name] = testCase
-		}
-	}
-
-	if len(focusedCases) > 0 {
-		testsToRun = focusedCases
-		focused = true
-	}
-	return testsToRun, focused
-}
-
-// TemplateTestCase is an individual template test.
-// Given and Expect values must be provided.
-// Fields in the expected object's metadata may be ignored
-// When run as part of a TemplateTestSuite, an individual case(s) may be focused.
-// This will exercise the individual test(s).
-// Note that the overall suite will fail (preventing focused tests from passing CI).
-type TemplateTestCase struct {
-	Given                TemplateTestGivens
-	Expect               TemplateTestExpectation
-	IgnoreMetadata       bool
-	IgnoreOwnerRefs      bool
-	IgnoreLabels         bool
-	IgnoreMetadataFields []string
-	Focus                bool
-}
-
-func (c *TemplateTestCase) Run() error {
-	expectedObject, err := c.Expect.getExpectedObject()
-	if err != nil {
-		return fmt.Errorf("failed to get expected object: %w", err)
-	}
-
-	actualObject, err := c.Given.getActualObject()
-	if errors.Is(err, yttNotFound) {
-		return fmt.Errorf("test requires ytt, but ytt was not found in path")
-	} else if err != nil {
-		return fmt.Errorf("failed to get actual object: %w", err)
-	}
-
-	c.stripIgnoredFields(expectedObject, actualObject)
-
-	if diff := cmp.Diff(expectedObject.Object, actualObject.Object); diff != "" {
-		return fmt.Errorf("expected does not equal actual: (-expected +actual):\n%s", diff)
-	}
-
-	return nil
-}
-
-func (c *TemplateTestCase) stripIgnoredFields(expected *unstructured.Unstructured, actual *unstructured.Unstructured) {
-	delete(expected.Object, "status")
-	delete(actual.Object, "status")
-
-	if c.IgnoreLabels {
-		expected.SetLabels(nil)
-		actual.SetLabels(nil)
-	}
-
-	if c.IgnoreMetadata {
-		delete(expected.Object, "metadata")
-		delete(actual.Object, "metadata")
-	}
-
-	var expectedMetadata, actualMetadata map[string]interface{}
-
-	if expected.Object["metadata"] != nil {
-		expectedMetadata = expected.Object["metadata"].(map[string]interface{})
-	}
-	if actual.Object["metadata"] != nil {
-		actualMetadata = actual.Object["metadata"].(map[string]interface{})
-	}
-
-	if c.IgnoreOwnerRefs {
-		delete(expectedMetadata, "ownerReferences")
-		delete(actualMetadata, "ownerReferences")
-	}
-
-	for _, field := range c.IgnoreMetadataFields {
-		delete(expectedMetadata, field)
-		delete(actualMetadata, field)
-	}
-}
-
-type TemplateTestExpectation interface {
-	getExpectedObject() (*unstructured.Unstructured, error)
-}
-
-type TemplateTestExpectedFile struct {
-	ExpectedFile string
-}
-
-type TemplateTestExpectedUnstructured struct {
-	ExpectedUnstructured *unstructured.Unstructured
-}
-
-func (e *TemplateTestExpectedUnstructured) getExpectedObject() (*unstructured.Unstructured, error) {
-	return e.ExpectedUnstructured, nil
-}
-
-type TemplateTestExpectedObject struct {
-	ExpectedObject client.Object
-}
-
-func (e *TemplateTestExpectedObject) getExpectedObject() (*unstructured.Unstructured, error) {
-	unstruct, err := runtime.DefaultUnstructuredConverter.ToUnstructured(e.ExpectedObject)
-	if err != nil {
-		return nil, fmt.Errorf("failed to convert template to unstructured: %w", err)
-	}
-
-	return &unstructured.Unstructured{Object: unstruct}, nil
-}
-
-func (e *TemplateTestExpectedFile) getExpectedObject() (*unstructured.Unstructured, error) {
-	expectedStampedObjectYaml, err := os.ReadFile(e.ExpectedFile)
-	if err != nil {
-		return nil, fmt.Errorf("could not read expected yaml: %w", err)
-	}
-
-	expectedJson, err := yaml.YAMLToJSON(expectedStampedObjectYaml)
-	if err != nil {
-		return nil, fmt.Errorf("convert yaml to json: %w", err)
-	}
-
-	expectedStampedObject := unstructured.Unstructured{}
-
-	if err = expectedStampedObject.UnmarshalJSON(expectedJson); err != nil {
-		return nil, fmt.Errorf("unmarshall json: %w", err)
-	}
-
-	return &expectedStampedObject, nil
 }
 
 // TemplateTestGivens must specify a template and a workload.
